test: move app/server util tests
Part of the test cleanup.
